requirements.txt
# ML Training & Inference
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Image Processing & Augmentation
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Dashboard
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
gdown>=4.7.0
